// web/liff/src/types/api/facility/models/CartResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 外部宿泊施設向け
 * 外部宿泊施設向けのふるマルAPIです。（公開エンドポイントについては「ふるマルAPI - 購入者向け」を参照してください）
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Product } from './Product';
import {
    ProductFromJSON,
    ProductFromJSONTyped,
    ProductToJSON,
} from './Product';
import type { Coordinator } from './Coordinator';
import {
    CoordinatorFromJSON,
    CoordinatorFromJSONTyped,
    CoordinatorToJSON,
} from './Coordinator';
import type { Cart } from './Cart';
import {
    CartFromJSON,
    CartFromJSONTyped,
    CartToJSON,
} from './Cart';

/**
 * 
 * @export
 * @interface CartResponse
 */
export interface CartResponse {
    /**
     * カート一覧
     * @type {Array<Cart>}
     * @memberof CartResponse
     */
    carts?: Array<Cart>;
    /**
     * コーディネータ一覧
     * @type {Array<Coordinator>}
     * @memberof CartResponse
     */
    coordinators?: Array<Coordinator>;
    /**
     * 商品一覧
     * @type {Array<Product>}
     * @memberof CartResponse
     */
    products?: Array<Product>;
}

/**
 * Check if a given object implements the CartResponse interface.
 */
export function instanceOfCartResponse(value: object): value is CartResponse {
    return true;
}

export function CartResponseFromJSON(json: any): CartResponse {
    return CartResponseFromJSONTyped(json, false);
}

export function CartResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CartResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'carts': json['carts'] == null ? undefined : ((json['carts'] as Array<any>).map(CartFromJSON)),
        'coordinators': json['coordinators'] == null ? undefined : ((json['coordinators'] as Array<any>).map(CoordinatorFromJSON)),
        'products': json['products'] == null ? undefined : ((json['products'] as Array<any>).map(ProductFromJSON)),
    };
}

export function CartResponseToJSON(value?: CartResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'carts': value['carts'] == null ? undefined : ((value['carts'] as Array<any>).map(CartToJSON)),
        'coordinators': value['coordinators'] == null ? undefined : ((value['coordinators'] as Array<any>).map(CoordinatorToJSON)),
        'products': value['products'] == null ? undefined : ((value['products'] as Array<any>).map(ProductToJSON)),
    };
}