// web/liff/src/types/api/facility/models/AddCartItemRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 外部宿泊施設向け
 * 外部宿泊施設向けのふるマルAPIです。（公開エンドポイントについては「ふるマルAPI - 購入者向け」を参照してください）
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AddCartItemRequest
 */
export interface AddCartItemRequest {
    /**
     * 商品ID
     * @type {string}
     * @memberof AddCartItemRequest
     */
    productId: string;
    /**
     * 数量
     * @type {number}
     * @memberof AddCartItemRequest
     */
    quantity?: number;
}

/**
 * Check if a given object implements the AddCartItemRequest interface.
 */
export function instanceOfAddCartItemRequest(value: object): value is AddCartItemRequest {
    if (!('productId' in value) || value['productId'] === undefined) return false;
    return true;
}

export function AddCartItemRequestFromJSON(json: any): AddCartItemRequest {
    return AddCartItemRequestFromJSONTyped(json, false);
}

export function AddCartItemRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddCartItemRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'productId': json['productId'],
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
    };
}

export function AddCartItemRequestToJSON(value?: AddCartItemRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'productId': value['productId'],
        'quantity': value['quantity'],
    };
}